hands-on-projects/project-08-data-pipeline/app/dashboard-ui/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
